refactor: move home page into vertical feature
